+++ src/table.cc
@@ -28,8 +28,8 @@ along with this program. If not, see http://www.gnu.org/licenses.
 
 // roulette
 #include "pch.hh"
-#include "field.hh"
 #include "table.hh"
+#include "field.hh"
 #include "sets.hh"
 
 namespace
@@ -57,7 +57,7 @@ namespace roulette
 
 		// create fields
 		EField temp;
-		for (size_t i = 0; i < static_cast<size_t>(EField::Dummy3); i++)
+		for (uint64 i = 0; i < static_cast<uint64>(EField::Dummy3); i++)
 		{
 			temp = static_cast<EField>(i);
 			m_fields.insert(make_pair(temp, new Field(temp, this)));
@@ -613,14 +613,14 @@ namespace roulette
 		{
 			error_handler(error("limit for this bet type not defined"));
 		}
-		unsigned current_bet = static_cast<unsigned>(get<0>(*chip));
+		uint16 current_bet = static_cast<uint16>(get<0>(*chip));
 
 		for (auto iter : chips)
 		{
 			// we need to compare points, since there are 4 splits or corners inside a field.
 			if (get<1>(*iter).equal(get<1>(*chip)))
 			{
-				current_bet += static_cast<unsigned>(get<0>(*iter));
+				current_bet += static_cast<uint16>(get<0>(*iter));
 			}
 		}
 
@@ -649,7 +649,7 @@ namespace roulette
 #pragma region
 #endif // _MSC_VER
 
-	void Table::on_signal_spin(unsigned result)
+	void Table::on_signal_spin(uint16 result)
 	{
 		for (auto iter : m_fields)
 		{
@@ -665,7 +665,7 @@ namespace roulette
 		}
 	}
 
-	void Table::set_table_max(const unsigned& limit /*= 0*/)
+	void Table::set_table_max(const uint32& limit /*= 0*/)
 	{
 		Gtk::Window* top_window = dynamic_cast<Gtk::Window*>(get_toplevel());
 
@@ -703,7 +703,7 @@ namespace roulette
 		else error_handler(error("ERROR: get_toplevel did not return a top level window"));
 	}
 
-	unsigned Table::get_limit(const EBet& name)
+	uint16 Table::get_limit(const EBet& name)
 	{
 		if ((m_maxiter = m_maximums.find(name)) == m_maximums.end())
 			error_handler(error("Table -> get_limit -> iterator out of range"));
@@ -711,7 +711,7 @@ namespace roulette
 		return m_maxiter->second;
 	}
 
-	unsigned Table::get_minimum(const EMinimum & minimum)
+	uint16 Table::get_minimum(const EMinimum & minimum)
 	{
 		auto iter = m_minimums.find(minimum);
 		if (iter == m_minimums.end())
@@ -721,7 +721,7 @@ namespace roulette
 		return iter->second;
 	}
 
-	void Table::set_minimum(const EMinimum& name, const unsigned& minimum)
+	void Table::set_minimum(const EMinimum& name, const uint16& minimum)
 	{
 		if (name == EMinimum::Table && (minimum < 0))
 			error_handler(error("Table -> set_minimum -> Table minimum less then 0"));
@@ -730,7 +730,7 @@ namespace roulette
 			error_handler(error("Table -> set_minimum -> Bet minimum less then 1"));
 	}
 
-	void Table::set_maximum(const EBet& name, const unsigned& limit)
+	void Table::set_maximum(const EBet& name, const uint16& limit)
 	{
 		(m_maxiter = m_maximums.find(name)) != m_maximums.end() ? m_maxiter->second = limit :
 			error_handler(error("Table -> set_maximum -> Iterator out of range"));
